web/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
annotated-types==0.6.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
crispy-bootstrap5==2023.10
cryptography==41.0.7
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==7.0.1
django-admin-list-filter-dropdown==1.0.3
django-allauth==0.58.2
django-autoslug==1.9.9
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap-v5==1.0.11
django-bootstrap4==24.1
django-bootstrap5==23.3
django-cities-light==3.10.0
django-crispy-forms==2.1
django-filter==23.5
django-init-admin-command==0.1.0
django-model-utils==4.3.1
django-multiselectfield==0.1.12
django-permissions-auditor==1.2.0
django-queryset-csv==1.1.0
django-table-sort==0.5.0
django-tables2==2.7.0
django==4.2.8
djangorestframework==3.15.2; python_version >= '3.8'
gunicorn==21.2.0
html5lib==1.1
idna==3.6
lxml==5.1.0
numpy==1.24.4
oauthlib==3.2.2
oscrypto==1.3.0
packaging==24.2; python_version >= '3.8'
pandas==2.0.3
pillow==10.2.0
progressbar2==4.2.0
psycopg2-binary==2.9.9
pycparser==2.21
pydantic-core==2.16.3
pydantic==2.6.2
pyhanko-certvalidator==0.26.3
pyhanko==0.21.0
pyjwt[crypto]==2.8.0
pypdf==4.1.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-gettext==5.0
python-utils==3.8.1
python3-openid==3.2.0
pytz==2023.3.post1
pyyaml==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests-oauthlib==1.3.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
swapper==1.3.0
tinycss2==1.2.1
tqdm==4.66.1
typing-extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
unicodecsv==0.14.1
unidecode==1.3.7
uritools==4.0.2
urllib3==2.1.0
webencodings==0.5.1
whitenoise==6.6.0
xhtml2pdf==0.2.15
xlsxwriter==3.2.0
